Risolti ulteriori bug dovuti all'aggiornamento di karate

+++ protocolli/modipa/testsuite/src/utils/get_info_transazione.js
@@ -9,7 +9,7 @@ function get_info_transazione(id_transazione, nome_colonna) {
 
     db_sleep_before_read = karate.properties['db_sleep_before_read']
     
-    java.lang.Thread.sleep(db_sleep_before_read)
+    java.lang.Thread.sleep(java.lang.Integer.valueOf(db_sleep_before_read))
     DbUtils = Java.type('org.openspcoop2.core.protocolli.modipa.testsuite.DbUtils')
     db = new DbUtils(govwayDbConfig)
     dbquery = "select "+nome_colonna+" from transazioni where id='"+id_transazione+"'"
